+++ testBalloon-framework-core/src/androidMain/kotlin/de/infix/testBalloon/framework/core/TestPlatform.android.kt
@@ -25,6 +25,7 @@ public object TestPlatformAndroid : TestPlatform {
         ?: instrumentationArguments?.getString(variableName)
 }
 
+@Deprecated("This function has no compelling use case in testing. Scheduled for removal in TestBalloon 0.8.")
 public actual fun dispatcherWithParallelism(parallelism: Int): CoroutineDispatcher =
     Dispatchers.IO.limitedParallelism(parallelism)