+++ testBalloon-framework-core/src/jsHostedMain/kotlin/de/infix/testBalloon/framework/core/TestPlatform.jsHosted.kt
@@ -22,6 +22,7 @@ public interface TestPlatformJsHosted : TestPlatform {
     override fun threadDisplayName(): String = "single"
 }
 
+@Deprecated("This function has no compelling use case in testing. Scheduled for removal in TestBalloon 0.8.")
 public actual fun dispatcherWithParallelism(parallelism: Int): CoroutineDispatcher =
     Dispatchers.Default // single-threaded on JS