+++ testBalloon-framework-core/src/nativeMain/kotlin/de/infix/testBalloon/framework/core/TestPlatform.native.kt
@@ -35,6 +35,7 @@ public object TestPlatformNative : TestPlatform {
     override fun environment(variableName: String): String? = getenv(variableName)?.toKString()
 }
 
+@Deprecated("This function has no compelling use case in testing. Scheduled for removal in TestBalloon 0.8.")
 public actual fun dispatcherWithParallelism(parallelism: Int): CoroutineDispatcher =
     Dispatchers.IO.limitedParallelism(parallelism)